elife module for elife specific formatting logic.
